random seed is a long int.

+++ platform_sdcc/platform.c
@@ -12,7 +12,7 @@ void cf_platform_init()
 
         {
                 long unsigned int cpctime = fw_kl_time_please();
-                unsigned int seed = cpctime ^ ( cpctime >> 16 );
+                long unsigned int seed = cpctime ^ ( cpctime >> 16 );
                 dbgvar_lu( 0, cpctime );
                 dbgvar_u( 0, seed );
                 srand( seed );
